test/mistClientService/requirements.txt
grpcio
googleapis-common-protos
protobuf==3.20.*

